backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
openai>=1.50.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
Pillow>=10.0.0,<12.0.0
easyocr>=1.7.0,<2.0.0
